Add files via upload
This is a brief discussion of how catching exceptions can help with file errors, we use it to understand what catching exceptions means and more specifically what errors are. The subtle difference between an error and an exception and how we can tell our program to tell us/warn if there are any errors during execution by offering ways on how we are trying to handle the errors that arise out of running our code.