Add files via upload
Lists, tuples, and dictionaries are used to store and organize the data in an efficient manner. They are also important in referencing the stored data of a long list of our information.Our discussion looks at the distinction between list, tuples, and dictionaries, while specifically applying the use of the required (zip, enumerate and items) methods for working with each.